TFKUBE-109: Unit test updated
