--- test/unittest/bamboo_test.go
@@ -1,1 +1,0 @@
-package unittest
